requirements.txt
Django>=4.2
django-environ
django-allauth
gunicorn  # Optional, for deployment
